refactor: update filter structure in read-model-store tests
- Changed filter format in read-model-store-in-memory.test.ts to use an array of filter objects
- Updated imports in event-bus.test.ts, event-handler.test.ts, query-bus.test.ts, and query-handler.test.ts for consistency
